INF-3203_Assignment/requirements.txt
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
torch==2.2.0
torchvision==0.17.0
matplotlib==3.8.3
scikit-image==0.22.0